redo the database, add login
