+++ front/src/App.css
@@ -7,7 +7,7 @@
   min-height: 10%;
   display: flex;
   flex-direction: row;
-  justify-content: space-evenly;
+  justify-content: space-between;
   font-size: calc(2rem);
   align-items: center;
 }
