+++ front/src/Greetings/Greetings.css
@@ -1,0 +1,3 @@
+.loginContainer > ul {
+  list-style-type: none;
+}
