--- front/src/Login/Login.css
@@ -1,3 +1,0 @@
-.loginContainer > ul {
-  list-style-type: none;
-}
