--- hasura/seeds/1603717726455_agency_seed.sql
@@ -1,10 +1,0 @@
-INSERT INTO public."Agency" (id, name) VALUES ('193743a5-ae7a-4fe2-9c15-0257c8eb3ca6', 'Paris');
-INSERT INTO public."Agency" (id, name) VALUES ('16ed885a-2c59-4d3c-959c-cdda64635cf2', 'Nantes');
-INSERT INTO public."Agency" (id, name) VALUES ('0be4a679-928b-4cc6-8a9d-f641608ad825', 'Singapore');
-INSERT INTO public."Agency" (id, name) VALUES ('5d9c9535-ea37-456d-ad90-bdd7b7370673', 'Bordeaux');
-INSERT INTO public."Agency" (id, name) VALUES ('5e1301b0-8e5d-4ba4-9673-d2b874c0b749', 'Brest');
-INSERT INTO public."Agency" (id, name) VALUES ('205cb3f5-f970-426c-a839-26928ab7a0d7', 'Montréal');
-INSERT INTO public."Agency" (id, name) VALUES ('fbb7ba59-d273-44b7-885b-813a33701441', 'Grenoble');
-INSERT INTO public."Agency" (id, name) VALUES ('529d3316-8d0c-4630-a7a0-6c7e6e083e3e', 'Lyon');
-INSERT INTO public."Agency" (id, name) VALUES ('37151f04-b718-401c-a057-4ee9e509038b', 'Rennes');
-INSERT INTO public."Agency" (id, name) VALUES ('1ccbb30b-93ab-472c-a51d-a16e2eb68641', 'Lille');
